patch from richard:
this patch fix an issue with stateless sip requests.

In some scenarios (depending of DNS replies), the request
is being terminated before the message is sent. This patch
fixes the issue. 

For stateless requests, if A/AAAA records for target of SRV record
is not present in addl (additional list) of NAPTR/SRV DNS reply, we
will explicitly query for those.



git-svn-id: https://svn.creytiv.com/re/trunk@4370 55cdf58e-e41d-0410-8799-97a9d43fd940